fix: enable DefaultParser in default config
If the DefaultParser is not included in the config file, we lose out on
its auto-discovery feature (i.e. Tika won't pick the best parser for the
uploaded file, and just depend on the hardcoded parsers that were
defined).

This commit re-enables the DefaultParser so that Tika can use the right
parsers as it sees fit, while still keeping the param overrides for the
TesseractOCRParser.